Create function to update display
The display is now updated when you click any of the number buttons on
the calculator.  This currently only selects the one number pressed and
displays it.

I also updated the display div to be an id instead of a class in the
HTML document and updated the other files accordingly.