Create Operator Button logic and update Display
Created the logic so that once an operator button is pressed on the
calculator, it will capture the digits as the first number variable and
store the operator button selected as the operator.

Now that it is storing the first number variable, I was able to go back
and update the display logic to reset itself after the next digit button
is pressed and start the display over.