Calculator Now Functions
The calculator can now be fully used.  The user first inputs one number,
then presses and operator, next inputs another number, and last they
push the "=" button and they will get the result of the calculation.

If the user has selected one operator and later selects another one, the
calculator will use the last selected operator.

Lastly, I added some code that checks if the result of the calculation
has a decimal, if so, it will round to two decimal places as to not
overflow the display.