Make Operator Button semi-transparent on click
When an operator button is clicked it will become transparent to let the
user know which operator they have selected.

It will be reverted to its original state if the user completes a
calculation or if the clear button is clicked.

If another operator is chosen while one is already selected, it will
revert the original operator to its previous state and make the new one
selected semi-transparent and use that one for the calculation.